Borrar Evento + Cambios en entidad Asistencia

+++ server/ejbModule/com/services/eventos/EventoBean.java
@@ -1,5 +1,6 @@
 package com.services.eventos;
 
+import java.time.LocalDateTime;
 import java.util.List;
 
 import javax.ejb.Stateless;
@@ -7,6 +8,7 @@ import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
 import javax.persistence.Persistence;
 import javax.persistence.Query;
+import javax.persistence.TypedQuery;
 
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
@@ -14,9 +16,11 @@ import org.hibernate.SessionFactory;
 import com.entities.Estudiante;
 import com.entities.Evento;
 import com.entities.Tutor;
+import com.enumerators.EnumAsistenciaEstado;
 
 @Stateless
 public class EventoBean implements EventoBeanRemote {
+	AsistenciaBean asistenciaBean = new AsistenciaBean();
 	
 	EntityManagerFactory emf = Persistence.createEntityManagerFactory("PDT-Server");
 	EntityManager em = emf.createEntityManager();
@@ -33,31 +37,49 @@ public class EventoBean implements EventoBeanRemote {
 		session.beginTransaction();
 		try {
 			em.persist(evento);
+			session.getTransaction().commit();
 			em.flush();			
 		} catch (Exception e) {
 			throw new Exception(e.getMessage());
 		}
-		session.getTransaction().commit();
 		session.close();
 	}
 
 	@Override
 	public void update(Evento evento) throws Exception {
+		session = factory.openSession();
+		session.beginTransaction();
 		try {
 			em.merge(evento);
+			session.getTransaction().commit();
 			em.flush();
 		} catch (Exception e) {
 			 throw new Exception("No se pudo actualizar el Evento");
 		}
+		session.close();
 	}
 
 	@Override
 	public void delete(Long id) throws Exception {
+		Evento evento = findById(id);
+		
+		// empezado, es true cuando el evento ya empezó
+		boolean empezado = evento.getFechaInicio().isBefore(LocalDateTime.now());
+		
+		// participaron, es true cuando el evento ya cuenta con estudiantes el estado TRUE (1) en el campo ASISTENCIA
+		boolean participaron = asistenciaBean.findByStatus(id, EnumAsistenciaEstado.ASISTENCIA).size() > 0;
+		
+		// Verifica que el evento no haya empezado y no tenga estudiantes participando (1)
+		if (empezado && participaron) {
+			System.out.println("No se pudo eliminar el evento. El evento ya inició y ya cuenta con participación de los estudiantes.");
+			throw new Exception("No se pudo eliminar el evento. El evento ya inició y ya cuenta con participación de los estudiantes.");
+		}
+		
 		try {
-			Evento evento = findById(id);
 			em.remove(evento);
-			em.flush();				
+			em.flush();
 		} catch (Exception e) {
+			e.printStackTrace();
 			throw new Exception("No se pudo eliminar el evento");
 		}
 	}
@@ -69,7 +91,14 @@ public class EventoBean implements EventoBeanRemote {
 
 	@Override
 	public List<Evento> findAll() {
-		Query query = em.createNamedQuery("Evento.findAll", Evento.class);
+		TypedQuery<Evento> query = em.createNamedQuery("Evento.findAll", Evento.class);
+		return query.getResultList();
+	}
+	
+	@Override
+	public List<Evento> findByTutor(Long idTutor) {
+		TypedQuery<Evento> query = em.createNamedQuery("Evento.findByTutor", Evento.class);
+		query.setParameter("id", idTutor);
 		return query.getResultList();
 	}
 	
@@ -81,5 +110,7 @@ public class EventoBean implements EventoBeanRemote {
 	public Estudiante getEstudiante(Long id){
 		return (Estudiante) em.find(Estudiante.class, id);
 	}
+
+	
 
 }
